template creation- moving BEK util implementation
commit 2
