+++ VMEncryption/main/test/test_bek_util.py
@@ -4,7 +4,8 @@ try:
 except ImportError:
     import mock  # python2
 
-from BekUtil import BekUtil, BekMissingException
+from BekUtil import BekUtil
+from IntefaceBekUtilImpl import BekMissingException,IntefaceBekUtilImpl
 from DiskUtil import DiskUtil
 from Common import CommonVariables
 from .console_logger import ConsoleLogger
@@ -26,7 +27,7 @@ class Test_Bek_Util(unittest.TestCase):
         disk_util_mock.get_mount_items.return_value = [{"src":"/dev/sdc1", "dest":"/nobek", "fs":"vfat"}, {"src":"/dev/sda1", "dest":"/", "fs":"ext4"}]
         bek_expected, fault_reason = bek_util.is_bek_volume_mounted_and_formatted()
         self.assertFalse(bek_expected)
-        self.assertEqual(fault_reason, bek_util.not_mounted_msg)
+        self.assertEqual(fault_reason, IntefaceBekUtilImpl.not_mounted_msg)
 
     @mock.patch('DiskUtil.DiskUtil', autospec=True)
     def test_is_bek_volume_mounted_and_formatted_wrong_fs(self, disk_util_mock):
@@ -34,7 +35,7 @@ class Test_Bek_Util(unittest.TestCase):
         disk_util_mock.get_mount_items.return_value = [{"src":"/dev/sdc1", "dest":"/mnt/azure_bek_disk", "fs":"wrongFS"}, {"src":"/dev/sda1", "dest":"/", "fs":"ext4"}]
         bek_expected, fault_reason = bek_util.is_bek_volume_mounted_and_formatted()
         self.assertFalse(bek_expected)
-        self.assertEqual(fault_reason, bek_util.wrong_fs_msg)
+        self.assertEqual(fault_reason, IntefaceBekUtilImpl.wrong_fs_msg)
 
     @mock.patch('os.path.exists')
     @mock.patch('DiskUtil.DiskUtil', autospec=True)
@@ -51,7 +52,7 @@ class Test_Bek_Util(unittest.TestCase):
         exists_mock.side_effect = [False, False]
         bek_attached, error_reason = bek_util.is_bek_disk_attached_and_partitioned()
         self.assertFalse(bek_attached)
-        self.assertEqual(error_reason, bek_util.bek_missing_msg)
+        self.assertEqual(error_reason, IntefaceBekUtilImpl.bek_missing_msg)
 
     @mock.patch('os.path.exists')
     @mock.patch('DiskUtil.DiskUtil', autospec=True)
@@ -60,7 +61,7 @@ class Test_Bek_Util(unittest.TestCase):
         exists_mock.side_effect = [True, False]
         bek_attached, error_reason = bek_util.is_bek_disk_attached_and_partitioned()
         self.assertFalse(bek_attached)
-        self.assertEqual(error_reason, bek_util.partition_missing_msg)
+        self.assertEqual(error_reason, IntefaceBekUtilImpl.partition_missing_msg)
 
     @mock.patch('os.path.exists')
     @mock.patch('DiskUtil.DiskUtil', autospec=True)
@@ -77,7 +78,7 @@ class Test_Bek_Util(unittest.TestCase):
         exists_mock.side_effect = [False, False]
         bek_attached, error_reason = bek_util.is_bek_disk_attached_and_partitioned()
         self.assertFalse(bek_attached)
-        self.assertEqual(error_reason, bek_util.bek_missing_msg)
+        self.assertEqual(error_reason, IntefaceBekUtilImpl.bek_missing_msg)
 
     @mock.patch('os.path.exists')
     @mock.patch('DiskUtil.DiskUtil', autospec=True)
@@ -86,6 +87,6 @@ class Test_Bek_Util(unittest.TestCase):
         exists_mock.side_effect = [False, True, False]
         bek_attached, error_reason = bek_util.is_bek_disk_attached_and_partitioned()
         self.assertFalse(bek_attached)
-        self.assertEqual(error_reason, bek_util.partition_missing_msg)
+        self.assertEqual(error_reason, IntefaceBekUtilImpl.partition_missing_msg)
